診療会計肥大化対応

+++ sql/4.4/orcadbcr_sryacct_sub_4.4.0-1.sql
@@ -200,7 +200,7 @@ or day3_31       > 0
 
 create table tbl_sryacct_sub(
 hospnum      integer         ,
-ptid         numeric(10,0)   ,
+ptid         bigint   ,
 zainum      integer          ,
 rennum      smallint         ,
 date        smallint         ,
